// types/models/eft/common/IGlobals.ts
export interface FenceLevel 
{ 
    SavageCooldownModifier: number
    ScavCaseTimeModifier: number
    PaidExitCostModifier: number
    BotFollowChance: number
    ScavEquipmentSpawnChanceModifier: number
    PriceModifier: number
    HostileBosses: boolean
    HostileScavs: boolean
    ScavAttackSupport: boolean
    ExfiltrationPriceModifier: number
    AvailableExits: number
}