// types/models/enums/ConfigTypes.ts
export enum ConfigTypes
    {
    AIRDROP = "aki-airdrop",
    BOT = "aki-bot",
    CORE = "aki-core",
    HEALTH = "aki-health",
    HIDEOUT = "aki-hideout",
    HTTP = "aki-http",
    IN_RAID = "aki-inraid",
    INSURANCE = "aki-insurance",
    INVENTORY = "aki-inventory",
    ITEM = "aki-item",
    LOCALE = "aki-locale",
    LOCATION = "aki-location",
    MATCH = "aki-match",
    PLAYERSCAV = "aki-playerscav",
    QUEST = "aki-quest",
    RAGFAIR = "aki-ragfair",
    REPAIR = "aki-repair",
    SCAVCASE = "aki-scavcase",
    TRADER = "aki-trader",
    WEATHER = "aki-weather",
    SEASONAL_EVENT = "aki-seasonalevents"
}